requirements/live.txt
Django==2.0.6
djangorestframework==3.8.2
pytz==2018.4
wincertstore==0.2
djangorestframework-jwt==1.11.0
